msgr: drop unnecessary (un)locking on queuing connection events
This used to be necessary because the pipe_lock was used when queueing
the pipe in the dispatch queue.  Now that is handled by IncomingQueue's
own lock, so these can be removed.

If we're locking, to eliminates a whoe category of weird races and
suble bugs due to losing pipe_lock momentarily.

Signed-off-by: Sage Weil <sage@inktank.com>